requirements.txt
Django>=5.2.8
django-ninja>=1.1.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
django-ninja-jwt>=5.4.0
django-cors-headers>=4.3.1
email-validator>=2.0.0

